# frontend/src/generated/openapi.yml
openapi: 3.0.3
info:
  title: Merism API
  version: 0.1.0
  description: AI-moderated qualitative research platform.
paths:
...
  /api/participants/{id}/:
    get:
      operationId: participants_retrieve
      description: |-
        Model viewset that filters queryset by the caller's current team.

        Usage::

            class StudyViewSet(TeamScopedModelViewSet):
                queryset = Study.objects.all()
                serializer_class = StudySerializer

        Override ``get_queryset()`` to further filter; the default implementation
        already restricts rows to ``team=self.get_team()`` if the model has a
        ``team`` FK.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: 标识此 参与者 的 UUID 字符串。
        required: true
      tags:
      - participants
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
          description: ''
    put:
      operationId: participants_update
      description: |-
        Model viewset that filters queryset by the caller's current team.

        Usage::

            class StudyViewSet(TeamScopedModelViewSet):
                queryset = Study.objects.all()
                serializer_class = StudySerializer

        Override ``get_queryset()`` to further filter; the default implementation
        already restricts rows to ``team=self.get_team()`` if the model has a
        ``team`` FK.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: 标识此 参与者 的 UUID 字符串。
        required: true
      tags:
      - participants
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Participant'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Participant'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Participant'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
          description: ''
    patch:
      operationId: participants_partial_update
      description: |-
        Model viewset that filters queryset by the caller's current team.

        Usage::

            class StudyViewSet(TeamScopedModelViewSet):
                queryset = Study.objects.all()
                serializer_class = StudySerializer

        Override ``get_queryset()`` to further filter; the default implementation
        already restricts rows to ``team=self.get_team()`` if the model has a
        ``team`` FK.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: 标识此 参与者 的 UUID 字符串。
        required: true
      tags:
      - participants
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedParticipant'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedParticipant'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedParticipant'
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Participant'
          description: ''
    delete:
      operationId: participants_destroy
      description: |-
        Model viewset that filters queryset by the caller's current team.

        Usage::

            class StudyViewSet(TeamScopedModelViewSet):
                queryset = Study.objects.all()
                serializer_class = StudySerializer

        Override ``get_queryset()`` to further filter; the default implementation
        already restricts rows to ``team=self.get_team()`` if the model has a
        ``team`` FK.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: 标识此 参与者 的 UUID 字符串。
        required: true
      tags:
      - participants
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    Participant:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        external_id:
          type: string
          maxLength: 200
        email:
          type: string
          format: email
          maxLength: 254
        name:
          type: string
          maxLength: 200
        attributes: {}
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
...
    PatchedParticipant:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        external_id:
          type: string
          maxLength: 200
        email:
          type: string
          format: email
          maxLength: 254
        name:
          type: string
          maxLength: 200
        attributes: {}
        created_at:
          type: string
          format: date-time
          readOnly: true
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid